feat: Move show logic inside AnniversaryAchievementModal
