+++ apps/web/src/components/GlobalCheckClaimStatus/index.tsx
@@ -1,15 +1,7 @@
-import { useEffect, useRef, useState } from 'react'
-import { useTranslation } from '@pancakeswap/localization'
-import { useModal, useToast } from '@pancakeswap/uikit'
 import { ChainId } from '@pancakeswap/sdk'
-import { useAccount } from 'wagmi'
-import { useRouter } from 'next/router'
-import { useActiveChainId } from 'hooks/useActiveChainId'
-import useCatchTxError from 'hooks/useCatchTxError'
-import { ToastDescriptionWithTx } from 'components/Toast'
 import useAccountActiveChain from 'hooks/useAccountActiveChain'
-import { useAnniversaryAchievementContract } from 'hooks/useContract'
 import AnniversaryAchievementModal from './AnniversaryAchievementModal'
+import V3AirdropModal from './V3AirdropModal'
 
 interface GlobalCheckClaimStatusProps {
   excludeLocations: string[]
@@ -34,62 +26,12 @@ const GlobalCheckClaimStatus: React.FC<React.PropsWithChildren<GlobalCheckClaimS
  */
 
 const GlobalCheckClaim: React.FC<React.PropsWithChildren<GlobalCheckClaimStatusProps>> = ({ excludeLocations }) => {
-  const { t } = useTranslation()
-  const { pathname } = useRouter()
-  const { address: account } = useAccount()
-  const { chainId } = useActiveChainId()
-  const hasDisplayedModal = useRef(false)
-  const contract = useAnniversaryAchievementContract({ chainId: ChainId.BSC })
-
-  const { toastError, toastSuccess } = useToast()
-  const { fetchWithCatchTxError } = useCatchTxError()
-
-  const [canClaimAnniversaryPoints, setCanClaimAnniversaryPoints] = useState(false)
-
-  const [onPresentAnniversaryModal] = useModal(
-    <AnniversaryAchievementModal
-      onClick={async () => {
-        try {
-          const receipt = await fetchWithCatchTxError(() => contract.write.claimAnniversaryPoints({ account, chainId }))
-
-          if (receipt?.status) {
-            toastSuccess(t('Success!'), <ToastDescriptionWithTx txHash={receipt.transactionHash} />)
-          }
-        } catch (error: any) {
-          const errorDescription = `${error.message} - ${error.data?.message}`
-          toastError(t('Failed to claim'), errorDescription)
-        }
-      }}
-    />,
+  return (
+    <>
+      <AnniversaryAchievementModal excludeLocations={excludeLocations} />
+      <V3AirdropModal />
+    </>
   )
-
-  // Check claim status
-  useEffect(() => {
-    const fetchClaimAnniversaryStatus = async () => {
-      const canClaimAnniversary = await contract.read.canClaim([account])
-      setCanClaimAnniversaryPoints(canClaimAnniversary)
-    }
-
-    if (account && chainId === ChainId.BSC) {
-      fetchClaimAnniversaryStatus()
-    }
-  }, [account, chainId, contract])
-
-  useEffect(() => {
-    const matchesSomeLocations = excludeLocations.some((location) => pathname.includes(location))
-
-    if (canClaimAnniversaryPoints && !matchesSomeLocations && !hasDisplayedModal.current) {
-      onPresentAnniversaryModal()
-      hasDisplayedModal.current = true
-    }
-  }, [pathname, excludeLocations, hasDisplayedModal, canClaimAnniversaryPoints, onPresentAnniversaryModal])
-
-  // Reset the check flag when account changes
-  useEffect(() => {
-    hasDisplayedModal.current = false
-  }, [account, hasDisplayedModal])
-
-  return null
 }
 
 export default GlobalCheckClaimStatus
